New Version
This uses a different algorithm for the prestige transmission, where
social teachers are selected for transmission with a probability
directly proportion to their prestigeIndex value + 1. This is different
from the code used in transmission.R which randomly selects the social
teachers, and the establish the probability of transmission as the
ratio between the social teachers prestige and the maximum possible
prestige (i.e. transmission always occurs when the social teacher has
the highest possible value at its prestigeIndex).